Hack to rename folders - step 2
