+++ src/walletSdk/Util/camelToSnakeCase.ts
